WindowServer: Starting a drag should forget the active input window
When we're in a drag, we're no longer concerned with streaming mouse
events to the window that initiated the drag, so just clear the active
input window pointer.

This fixes an issue where you'd have to click once after drag and drop
in order to "release" the mouse from the active input window.