Help: Directly launch an app from its help page
Through links in the help page, the user can directly launch the app
whose help page is currently viewed. The idea for this feature came up
in the discussion of #11557
(https://github.com/SerenityOS/serenity/pull/11557#issuecomment-1004830728
). The intention is that the user can simply open the app they are
currently trying to understand, and play around with it, learn by doing,
or follow along with any guide that may be present in the help page. It
feels very great :^)