Ports/gnuplot: Explicitly specify which optional dependencies to enable
This change explicitly enables support for `libgd` and the lua
terminal in `gnuplot`, while explicitly disabling `cairo` terminals,
which don't work with our `cairo` port. Adding `libgd` as a
dependency requires us to manually link against the dependencies of
`libgd`, as this is not done automatically by the configure script.

This fixes an issue where building `gnuplot` would fail if `libgd`
was already installed.