Ladybird: Do not include comment start/end sequence in its editable text
Currently, when editing a comment, the `<!--` and `-->` start and end
sequences would be included in the generated <input> field. This would
result in including that text in the updated comment text. So, for
example, in a comment such as:

    <!-- foo -->

Changing "foo" to "bar" would result in the comment:

    <!--<!-- bar -->-->

And this would repeatedly nest for each edit.