Kernel/Net: Clean up the ICMP handler
This change modernizes various aspects of handle_icmp.

A few dbgln's got turned into dbgln_if's, in accordance with the
behavior throughout the rest of the file.

Furthermore, we revert a part of commit ad73adef5d2f8 which introduced
inconsistency between struct names. There's no ICMPv4, only ICMP and
ICMPv6. Thus, it makes more sense to revert the ICMPv4Type back to
ICMPType.

We also remove the FIXME about the TTL - as per RFC1700, TTL of 64 is
a recommended default, so the current behavior can stay.