WebContent: Implement the WebDriver concept of current browsing context
When we create a WebDriverConnection object, we currently hand it the
page client for which it was opened, and perform all actions on that
client. However, some WebDriver endpoints change the browsing context
(and therefore page client) on which future commands should be executed.
For example, the switch-frame endpoint will switch the current browsing
context to a frame/iframe context.

This patch implements the current browsing context (and current top-
level browsing context) concepts. They are initialized to that of the
original page. Most of this patch is making sure we execute actions on
the correct context.

(cherry picked from commit c4da0228647276821f99887014ed1dae3cea8061;
amended to fix conflicts due to serenity not having the renderer-side
backing stores from LadybirdBrowser/ladybird#294 and due to serenity
not having the redundant logical op introduced in
LadybirdBrowser/ladybird#1021)