LibWeb/Fetch: Do not clone stored responses
Reading the RFC9111 spec makes it clear that the stored response was
not intended to be cloned. This is because there is a "clone response"
operation that is used in other places, but never for stored responses.

(cherry picked from commit afe74afa9e61569d35797d6ada5d54f0c22da412)