CMake: Explicitly mark LibC TARGET custom commands as POST_BUILD
This avoids a warning for CMake policy CMP0175, which is new with CMake
3.31. The warning is as follows:

Exactly one of PRE_BUILD, PRE_LINK, or POST_BUILD must be given.
Assuming POST_BUILD to preserve backward compatibility.

To preserve behavior, mark these as POST_BUILD. It might not be the
ideal way to have the custom commands run, but it is behavior-preserving