LibGfx/JPEG2000: Rewrite segments_for_all_layers()
The job of this function is to gather segment data for segments
that are spread across multiple layers, and to return a vector
with one span for each final segment.

For segments that do not need relocating, it can return a span to the
original input data, without copying.

Previously, this function decided what to do based on the code
block style.

But now that we store the index of each segment in a layer, we can
instead build a segment id -> segment parts list. With that,
it is easy to copy the data of segments with multiple
parts into a buffer and then making a vector with spans for all
complete segment bits.

This has the advantage that it will work for the selective arithmetic
decoder bypass code block style; both combined with the terminate
each coding pass style and standalone.

Also allows removing a special case from a loop.

No behavior change.