Kernel/USB: Remove the HID boot protocol mouse driver
Now that we support the HID report protocol, we don't need this mouse
boot protocol driver anymore. The report protocol is required to be
supported on all USB HID devices, while the boot protocol is optional.

Even if some devices claim they support the boot protocol, they
sometimes appears to have a broken implementation of it, as it was
probably never tested because most operating systems only use the
report protocol.