Tests/JPEG2000: Add a test for overflowing the tile part index field
The progression order tests accidentally tested for that previously.
I fixed that now, but it seems nice to have a test for this, given
it's easy to create.

I created the input file by running:

    ../openjpeg/build/bin/opj_compress -i ref.bmp \
        -o output0-many.jp2 \
        -n 2 -b 4,16 -c '[16,32]' -t 39,53 -r 20,10,1 -TP P -p LRCP \
        -C "opj_compress <all flags from previous line>"

This uses progression order 0 and splits tile parts on the
fastest-changing part of this order (P). opj_compress then creates
hundreds of tile parts.  According to Table A.5, the tile-part index
TPsot is supposed to be in [0, 254], so these images are technically
invalid per spec -- and jasper, grok, and even opj_decompress do refuse
to decode them.

Preview.app and Photoshop show them no problem though, so we'll
want to be able to decode files with >= 255 tile parts at some point.