LibPDF: Clamp shading types 4-7 with functions to correct bounds
Plumb component bounds around and use them for clamping,
which fixes three FIXMEs.

This is only needed because due to aliasing, the barycentric
coordinates in GouraudPaintStyle sometimes create coefficents
slightly below 0 or above 1. We used to just clamp function
bounds to [0, 1], which does the wrong thing if those aren't
the function bounds.