LibPDF: Reject graphics state dicts with /Font entry
I think the correct implementation of this is just

    auto array = MUST(dict->get_array(m_document, CommonNames::Font));
    TRY(set_font(TRY(array->get_dict_at(m_document, 0)),
                 array->at(1).to_float()));

but there are zero PDF files using this in my 2000-file test set,
so I can't check if it works. If I put in an error for it, maybe
someone will notice when they hit it on some file, and then we
can put in the actual code then.