LibGfx/JBIG2: Fix incorrect composition with negative coordinates
If a pattern was drawn with a negative x but was still wide enough
to reach into the frame (and analogously for y), we both increased
the start x that the pattern was drawn at and set the origin x to 0.
But then, `position.x() + x` didn't start at 0 but at the absolute
value of the negative offset.

Instead, only increase the start offset in the pattern, but don't
change the negative x coordinate. This way, `position.x() + x`
will start at 0.

Improves e.g. the appearance of the upper left image edges of
200-20-45.jb2.