Tests/JBIG2: Add test for negative SBDSOFFSET
This would've caught the JBIG2 regression introduced by
the first commit in #25805.

I created this file as described in the commit that added
bitmap-symbol.jbig2 (3a50cadddf817), but added `-txt -Param -numPich -5`
right before EOF.

I had to teach `jbig2` that sbdsoffset can be negative, see the PR
with this commit for my local patch.