LibGfx/CCITT: Align after decoding group 4 lines, not before it
The bitstream is aligned before the first line.

0000871.pdf has `/EndOfLine` set (implicitly), and also an explicit
`/EncodedByteAlign`. It looks like it is aligned to the next byte
before the EOFB.

The PDF spec is not very explicit on if `/EncodedByteAlign` should
cause alignment after the last scanline before EOFB, but with a
future change to require EOFB it unbreaks 0000871.pdf and doesn't
seem to cause any new problems, so let's go with this for now.