LibGfx/JBIG2: Fix off-by-one in max stripe height computation
Rename field in page information segment to maximum_stripe_size
to match its name in the spec.

7.4.8.6 Page striping information:

> the maximum size of each stripe (the distance between an end of stripe
> segment's end row and the end row of the previous end of stripe
> segment, or 0 in the case of the first end of stripe segment) must be
> no more than the page's maximum stripe size.

That means that the maximum stripe _height_ is one more than the
maximum strip size, so update the code accordingly.

This lets us decode 0001341.pdf page 6, which has a page information
segment with height=1408 and max_stripe_size=1407.