LibGfx/JBIG2+Tests: Add a very basic JBIG2Writer
For now, it can encode any image as long as it's completely white.
It will learn to encode more images in the future.

The motivation for this writer is to be able to create jbig2 images
to be used as test cases. I created the existing images in
Tests/LibGfx/test-inputs/jbig2/ by running the reference `jbig2`
tool, but it's missing many features, and the features that are there
are buggy and I had to patch them up. (And the signal test images
don't decode in Preview.app.) Maybe it's easier to just write our
own encoder.

It might even come in somewhat handy if we ever add a PDF writer
and want to store bilevel images in a PDF.