LibGfx/JBIG2: Teach JBIG2Writer to write generic immediate regions
JBIG2Writer can now write arithmetically-coded image data.
This lets it write basically all images; all remaining stuff is
just for better compression (and for adding more test files for
the decoder, which is the main motivation here, really).

generic_region_encoding_procedure() is very similar to
generic_region_decoding_procedure(), only the loop at the end is
different (to compress data instead of to decompress it). Maybe
we can find a way to share more code in the future. Quite a bit
of support code is already shared, though.

I also need to come up with a way to pass many, many options to
the encoder. For now, this just hardcodes all the options to
somewhat useful defaults.

Likewise, we also need a mechanism to manually pick which segments
to write. A fully white or fully black image does not need an
immediate generic region, for example. (But for normal `image`
use, we should also automatically generate a schedule that makes
sense for a given image.)