LibGfx/JBIG2: Re-implement JBIG2Writer::encode()
Now that encode_with_explicit_data() is expressive enough that
encode() can be implemented in terms of it, do so.

JBIG2EncoderOptions is now empty. If we still don't have a use
for it once we can write symbol / text region segments, we should
remove it.

This now has a few too many magic numbers, which we should clean
up later.

No behavior change.