Tests/LibGfx: Make JBIG2 reference corner tests json-based
This uses the numbers from the third commit in #23810 (67875ecb6fdfe).
But apparently I pasted the `topright` numbers incorrectly. The numbers
in the json file match what was in the actual file. (The files would
look off otherwise, in all viewers I tried.)

The three json files are identical except for the `reference_corner`
parameter.