Tests/LibGfx: Add JBIG2 test with > 4 referred-to segments
Up to four referred-to segments use a single byte to store referred-to
segment count and retained bits. Five or more referred-to segments
cause a u32 for the count to be written (with the topmost 3 bits masked
out), and then one bit per referred-to segment.

This adds a test case for that "five or more" case.

There's a few places where this can happen, but the most natural seems
like a text segment referring to a bunch of symbol segments, so the
test is that.

The json file is basically bitmap-symbol-symhuff-texthuff.json, but
with huffman coding disabled, superfluous flags removed, and with one
dedicated symbol region per symbol.

This jbig2 file, when run through Meta/jbig2_to_pdf.py, decodes fine
in Preview.app and Acrobat Reader (I only checked on mobile).
PDFium, pdf.js, and mutool all can't decode it.