LibGfx/ICC+icc: Add --stdin-u8-to-pcs / --stdin-u8-from-pcs flags
This allows inspecting profile conversion intermediate results,
and running things like:

    echo '255, 0, 0' |
        icc profile1.icc --stdin-u8-to-pcs |
        icc profile2.icc --stdin-u8-from-pcs |
        icc profile2.icc --stdin-u8-to-pcs |
        icc profile1.icc --stdin-u8-from-pcs

The output of --stdin-u8-to-pcs prints data in the profile
connection space, which is either XYZ or LAB:

    % echo '255, 255, 255' | \
         Build/lagom/bin/icc -n sRGB --stdin-u8-to-pcs
    pcsxyz(0.9642, 0.99999994, 0.8249001)

    % echo '255, 255, 255, 128' | \
         Build/lagom/bin/icc \
           Build/lagom/Root/res/icc/Adobe/CMYK/USWebCoatedSWOP.icc \
           --stdin-u8-to-pcs
    pcslab(6.606001, 2.4342346, -1.8503876)

Which one of the two is printed is a property of the used .icc file.

--stdin-u8-from-pcs parses these two formats and prints bytes in
the data color space of the passed-in color profile.

This is useful for converting single colors from one profile to
another, and for inspecting internal LibGfx state (the PCS values).

As a comment in the code points out, the PCS->data conversion is
a tiny bit fishy as it also depends on the illuminant of the
_source_ profile. To be 100% correct, I think (but I'm not sure)
that we'd also have to write the illuminant to the output
(`pcsxyz(a, b, c, illuminant=d, e, f)` or something). But the V4
ICC spec mandates that the illuminant is the D50 illuminant, and
for most V2 profiles that's true too. So this takes a shortcut,
at least for now. Since ICC::Profile rejects anything that's not
using a D50 illuminant with a diag, it should be fine in practice.