LibGfx/ICC: Implement converting from PCS to LutBToATagData with clut
Similar to what #26452 did for Lut8TagData and #26462 for Lut16TagData.

With this, we can convert from PCS to all tag types :^)

Which means our ICC implementation is now fairly complete! We don't
support DeviceLink or Abstract profile types (both don't appear in
images; both are arguably useful in some not-embedded-in-images
scenarios), nor the optiona BToD / DToB tags, and are missing
some other minor things (like grayscale profiles -- but not much
is missing here). But overall, this is now a fairly complete (if slow)
basic ICC implementation :^)