LibGfx/JBIG2Writer: Add support for writing halftone skip patterns
Skip patterns are really only interesting for halftone regions with
a rotated grid, which we can't do yet. So the skip pattern will be
all "don't skip" entries everywhere for now.

But once we do add support for non-identity grids, this will hopefully
just work. And even with an identity grid, this already produces a
useful test file.