Tests/LibGfx: Add JBIG2 test for inner halftone region composition
Halftone regions allow setting a background color, and a combination
operator that's used when drawing pattern instances into the halftone
region.  (The final output of the halftone region is then composited
onto the page with the region's *external* composition operator.)
This adds tests for that.

Testing halftone combination operators requires overlapping tiles.
Normally, halftone regions have partially overlapping tiles, but for
tests it's easier if all tiles in a halftone region overlap completely.
To achieve this, this sets the grid offset vector to (0, 0) for all
halftone regions in the file.

Otherwise, this follows the same strategy as the inner text region tests
in #26418, amended with the extended tests for "or" and "and" in #26501
(and with "replace", since halftone regions allow this for the inner
combination operator, and without the double invert for "and" and
"xnor"):

To test for 'or', paint a region (the upper eye) twice in the same
location, on a white background. Paint a white region on top.

To test for 'xor', paint a region (the lower eye) once with the correct
data, and paint incorrect data twice at the same location, on a
white background.

To test for 'replace', draw a region (the nose) first in inverted, and
then again in uninverted.

To test for 'and', paint a region (the upper half of the mouth) twice,
on a black background. Draw a completely black region on top.

To test for 'xnor', paint an region (the lower half of the mouth) once
with the correct data, and paint incorrect data twice at the same
location, on a black background.

While here, also remove an outdated comment about missing tests for
halftone regions with rotated grids: They were added in #26496.