LibGfx/JBIG2: Handle text regions, symbol dictionaries without symbols
As far as I can tell, nothing in the spec forbids symbol dictionaries to
have 0 symbols, or text regions to not use any symbols.

Section 7.3.1 even explicitly says that text regions 'may refer to any
number of segments of type "symbol dictionary"', which sounds like it
includes 0.

The spec does say that ceil(log2(num_symbols)) should be computed
though, which is undefined if num_symbols is 0. But ceil(log2(1)) is
0, and using 0 for ceil(log2(0)) feels like what the spec intends.
Furthermore, this seems to be the behavior of other jbig2 decoders.

ceil_log2(x) is 0 for x == 0, so just switching from ceil(log(x))
to AK::ceil_log(x) has the effect that we want.

The test has one text region referring to a symbol dictionary without
symbols, and a text region not referring to any symbol dictionaries,
to make sure that those are drawn without exceptions.
Finally, it has a generic region with the actual image data on top.